pyproject.toml
[build-system]
requires = ["hatchling", "hatch-build-scripts"]
build-backend = "hatchling.build"

[project]
name = "telegram-bot-wrapper"
version = "0.1.0"
description = "Telegram Bot Wrapper"
requires-python = "==3.13.*"

dependencies = [
    "httpx==0.27.1",
    "pydantic-settings==2.5.2",
    "python-telegram-bot==22.5",
    "python-telegram-bot[job-queue]==22.5",
    "pytz==2024.1"
]

[tool.hatch.build.targets.wheel]
only-include = [
    "src/bot",
]

# Включаем py.typed и все .pyi файлы
[tool.hatch.build.targets.wheel.force-include]
"src/bot/py.typed" = "ikemurami/telegram/py.typed"

[tool.hatch.build.targets.wheel.sources]
"src/bot" = "ikemurami/telegram"
